Struct google_qpxexpress1::FareInfo

pub struct FareInfo {
    pub basis_code: Option<String>,
    pub origin: Option<String>,
    pub kind: Option<String>,
    pub carrier: Option<String>,
    pub destination: Option<String>,
    pub id: Option<String>,
    pub private: Option<bool>,
}

Complete information about a fare used in the solution to a low-fare search query. In the airline industry a fare is a price an airline charges for one-way travel between two points. A fare typically contains a carrier code, two city codes, a price, and a fare basis. (A fare basis is a one-to-eight character alphanumeric code used to identify a fare.)

This type is not used in any activity, and only used as part of another schema.

Fields

basis_code

no description provided

origin

The city code of the city the trip begins at.

kind

Identifies this as a fare object. Value: the fixed string qpxexpress#fareInfo.

carrier

The carrier of the aircraft or other vehicle commuting between two points.

destination

The city code of the city the trip ends at.

id

A unique identifier of the fare.

private

Whether this is a private fare, for example one offered only to select customers rather than the general public.