Struct google_qpxexpress1::FreeBaggageAllowance

pub struct FreeBaggageAllowance {
    pub kilos_per_piece: Option<i32>,
    pub bag_descriptor: Option<Vec<BagDescriptor>>,
    pub pounds: Option<i32>,
    pub pieces: Option<i32>,
    pub kind: Option<String>,
    pub kilos: Option<i32>,
}

Information about free baggage allowed on one segment of a trip.

This type is not used in any activity, and only used as part of another schema.

Fields

kilos_per_piece

The maximum number of kilos any one piece of baggage may weigh.

bag_descriptor

A representation of a type of bag, such as an ATPCo subcode, Commercial Name, or other description.

pounds

The number of pounds of free baggage allowed.

pieces

The number of free pieces of baggage allowed.

kind

Identifies this as free baggage object, allowed on one segment of a trip. Value: the fixed string qpxexpress#freeBaggageAllowance.

kilos

The maximum number of kilos all the free baggage together may weigh.