Struct google_qpxexpress1::LegInfo

pub struct LegInfo {
    pub origin: Option<String>,
    pub origin_terminal: Option<String>,
    pub mileage: Option<i32>,
    pub secure: Option<bool>,
    pub connection_duration: Option<i32>,
    pub departure_time: Option<String>,
    pub aircraft: Option<String>,
    pub on_time_performance: Option<i32>,
    pub arrival_time: Option<String>,
    pub duration: Option<i32>,
    pub destination: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub destination_terminal: Option<String>,
    pub change_plane: Option<bool>,
    pub operating_disclosure: Option<String>,
    pub meal: Option<String>,
}

Information about a leg. (A leg is the smallest unit of travel, in the case of a flight a takeoff immediately followed by a landing at two set points on a particular carrier with a particular flight number.)

This type is not used in any activity, and only used as part of another schema.

Fields

origin

The leg origin as a city and airport.

origin_terminal

The terminal the flight is scheduled to depart from.

mileage

The number of miles in this leg.

secure

Whether passenger information must be furnished to the United States Transportation Security Administration (TSA) prior to departure.

connection_duration

Duration of a connection following this leg, in minutes.

departure_time

The scheduled departure time of the leg, local to the point of departure.

aircraft

The aircraft (or bus, ferry, railcar, etc) travelling between the two points of this leg.

on_time_performance

In percent, the published on time performance on this leg.

arrival_time

The scheduled time of arrival at the destination of the leg, local to the point of arrival.

duration

The scheduled travelling time from the origin to the destination.

destination

The leg destination as a city and airport.

id

An identifier that uniquely identifies this leg in the solution.

kind

Identifies this as a leg object. A leg is the smallest unit of travel, in the case of a flight a takeoff immediately followed by a landing at two set points on a particular carrier with a particular flight number. Value: the fixed string qpxexpress#legInfo.

destination_terminal

The terminal the flight is scheduled to arrive at.

change_plane

Whether you have to change planes following this leg. Only applies to the next leg.

operating_disclosure

Department of Transportation disclosure information on the actual operator of a flight in a code share. (A code share refers to a marketing agreement between two carriers, where one carrier will list in its schedules (and take bookings for) flights that are actually operated by another carrier.)

meal

A simple, general description of the meal(s) served on the flight, for example: "Hot meal".