Struct google_qpxexpress1::PassengerCounts

pub struct PassengerCounts {
    pub infant_in_lap_count: Option<i32>,
    pub kind: Option<String>,
    pub infant_in_seat_count: Option<i32>,
    pub adult_count: Option<i32>,
    pub senior_count: Option<i32>,
    pub child_count: Option<i32>,
}

The number and type of passengers. Unfortunately the definition of an infant, child, adult, and senior citizen varies across carriers and reservation systems.

This type is not used in any activity, and only used as part of another schema.

Fields

infant_in_lap_count

The number of passengers that are infants travelling in the lap of an adult.

kind

Identifies this as a passenger count object, representing the number of passengers. Value: the fixed string qpxexpress#passengerCounts.

infant_in_seat_count

The number of passengers that are infants each assigned a seat.

adult_count

The number of passengers that are adults.

senior_count

The number of passengers that are senior citizens.

child_count

The number of passengers that are children.