Struct google_qpxexpress1::SliceInfo

pub struct SliceInfo {
    pub duration: Option<i32>,
    pub kind: Option<String>,
    pub segment: Option<Vec<SegmentInfo>>,
}

Information about a slice. A slice represents a traveller's intent, the portion of a low-fare search corresponding to a traveler's request to get between two points. One-way journeys are generally expressed using 1 slice, round-trips using 2. For example, if a traveler specifies the following trip in a user interface: | Origin | Destination | Departure Date | | BOS | LAX | March 10, 2007 | | LAX | SYD | March 17, 2007 | | SYD | BOS | March 22, 2007 | then this is a three slice trip.

This type is not used in any activity, and only used as part of another schema.

Fields

duration

The duration of the slice in minutes.

kind

Identifies this as a slice object. A slice represents a traveller's intent, the portion of a low-fare search corresponding to a traveler's request to get between two points. One-way journeys are generally expressed using 1 slice, round-trips using 2. Value: the fixed string qpxexpress#sliceInfo.

segment

The segment(s) constituting the slice.