Struct google_qpxexpress1::PricingInfo

pub struct PricingInfo {
    pub fare: Option<Vec<FareInfo>>,
    pub fare_calculation: Option<String>,
    pub kind: Option<String>,
    pub segment_pricing: Option<Vec<SegmentPricing>>,
    pub sale_total: Option<String>,
    pub ptc: Option<String>,
    pub passengers: Option<PassengerCounts>,
    pub tax: Option<Vec<TaxInfo>>,
    pub sale_fare_total: Option<String>,
    pub base_fare_total: Option<String>,
    pub refundable: Option<bool>,
    pub sale_tax_total: Option<String>,
    pub latest_ticketing_time: Option<String>,
}

The price of one or more travel segments. The currency used to purchase tickets is usually determined by the sale/ticketing city or the sale/ticketing country, unless none are specified, in which case it defaults to that of the journey origin country.

This type is not used in any activity, and only used as part of another schema.

Fields
fare: Option<Vec<FareInfo>>

The fare used to price one or more segments.

fare_calculation: Option<String>

The horizontal fare calculation. This is a field on a ticket that displays all of the relevant items that go into the calculation of the fare.

kind: Option<String>

Identifies this as a pricing object, representing the price of one or more travel segments. Value: the fixed string qpxexpress#pricingInfo.

segment_pricing: Option<Vec<SegmentPricing>>

The per-segment price and baggage information.

sale_total: Option<String>

Total per-passenger price (fare and tax) in the sale or equivalent currency.

ptc: Option<String>

The passenger type code for this pricing. An alphanumeric code used by a carrier to restrict fares to certain categories of passenger. For instance, a fare might be valid only for senior citizens.

passengers: Option<PassengerCounts>

The number of passengers to which this price applies.

tax: Option<Vec<TaxInfo>>

The taxes used to calculate the tax total per ticket.

sale_fare_total: Option<String>

The total fare in the sale or equivalent currency.

base_fare_total: Option<String>

The total fare in the base fare currency (the currency of the country of origin). This element is only present when the sales currency and the currency of the country of commencement are different.

refundable: Option<bool>

Whether the fares on this pricing are refundable.

sale_tax_total: Option<String>

The taxes in the sale or equivalent currency.

latest_ticketing_time: Option<String>

The latest ticketing time for this pricing assuming the reservation occurs at ticketing time and there is no change in fares/rules. The time is local to the point of sale (POS).