Struct google_qpxexpress1::SegmentInfo

pub struct SegmentInfo {
    pub kind: Option<String>,
    pub booking_code_count: Option<i32>,
    pub flight: Option<FlightInfo>,
    pub leg: Option<Vec<LegInfo>>,
    pub subject_to_government_approval: Option<bool>,
    pub connection_duration: Option<i32>,
    pub booking_code: Option<String>,
    pub duration: Option<i32>,
    pub id: Option<String>,
    pub cabin: Option<String>,
    pub married_segment_group: Option<String>,
}

Details of a segment of a flight; a segment is one or more consecutive legs on the same flight. For example a hypothetical flight ZZ001, from DFW to OGG, would have one segment with two legs: DFW to HNL (leg 1), HNL to OGG (leg 2), and DFW to OGG (legs 1 and 2).

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Identifies this as a segment object. A segment is one or more consecutive legs on the same flight. For example a hypothetical flight ZZ001, from DFW to OGG, could have one segment with two legs: DFW to HNL (leg 1), HNL to OGG (leg 2). Value: the fixed string qpxexpress#segmentInfo.

booking_code_count

The number of seats available in this booking code on this segment.

flight

The flight this is a segment of.

leg

The legs composing this segment.

subject_to_government_approval

Whether the operation of this segment remains subject to government approval.

connection_duration

In minutes, the duration of the connection following this segment.

booking_code

The booking code or class for this segment.

duration

The duration of the flight segment in minutes.

id

An id uniquely identifying the segment in the solution.

cabin

The cabin booked for this segment.

married_segment_group

The solution-based index of a segment in a married segment group. Married segments can only be booked together. For example, an airline might report a certain booking code as sold out from Boston to Pittsburgh, but as available as part of two married segments Boston to Chicago connecting through Pittsburgh. For example content of this field, consider the round-trip flight ZZ1 PHX-PHL ZZ2 PHL-CLT ZZ3 CLT-PHX. This has three segments, with the two outbound ones (ZZ1 ZZ2) married. In this case, the two outbound segments belong to married segment group 0, and the return segment belongs to married segment group 1.