Struct google_qpxexpress1::TripOptionsRequest

pub struct TripOptionsRequest {
    pub passengers: Option<PassengerCounts>,
    pub ticketing_country: Option<i64>,
    pub slice: Option<Vec<SliceInput>>,
    pub solutions: Option<i32>,
    pub refundable: Option<bool>,
    pub sale_country: Option<i64>,
    pub max_price: Option<String>,
}

A QPX Express search request, which will yield one or more solutions.

This type is not used in any activity, and only used as part of another schema.

Fields
passengers: Option<PassengerCounts>

Counts for each passenger type in the request.

ticketing_country: Option<i64>

IATA country code representing the point of ticketing.

slice: Option<Vec<SliceInput>>

The slices that make up the itinerary of this trip. A slice represents a traveler's intent, the portion of a low-fare search corresponding to a traveler's request to get between two points. One-way journeys are generally expressed using one slice, round-trips using two. An example of a one slice trip with three segments might be BOS-SYD, SYD-LAX, LAX-BOS if the traveler only stopped in SYD and LAX just long enough to change planes.

solutions: Option<i32>

The number of solutions to return, maximum 500.

refundable: Option<bool>

Return only solutions with refundable fares.

sale_country: Option<i64>

IATA country code representing the point of sale. This determines the "equivalent amount paid" currency for the ticket.

max_price: Option<String>

Do not return solutions that cost more than this price. The alphabetical part of the price is in ISO 4217. The format, in regex, is [A-Z]{3}\d+(.\d+)? Example: $102.07